Identifying and handling negative controls in array
Update to multinorm function to test for potential negative controls used in the array experiment. Briefly, if >90% of protein measurement results for a sample (column) are below the overall median of the dataframe, these samples are defined and shown to the user with column names. The user selects whether to exclude these samples from normalization. These samples are ultimately re-added to the output dataframe without any alteration.

Also, minor improvements to data handling, protein-based normalization calculation, and output generation.