Cut-off addition and hit summary corrections
A new column is added to summaries (implied_cutoff) so that this can be used by the user. It is called implied since this hit detection method does not generate a true threshold.

Summary values for hits now show raw values including the control group (may update to also apply this logic to the fc values).